Move bench-exchange to local cluster
