+++ local_cluster/src/tests.rs
@@ -1,3 +1,4 @@
+mod bench_exchange;
 mod bench_tps;
 mod local_cluster;
 mod replicator;
